* added layer mask to Node, Camera and Light for respective culling * added export of camera viewport * using 64bit sorting keys

+++ Wire/Engine/WireCullerSorting.h
@@ -31,7 +31,7 @@ public:
 		const Vector3F& rPosition = Vector3F::ZERO);
 
 protected:
-	UInt GetKey(RenderObject* pRenderObject, const Vector3F& rPosition);
+	UInt64 GetKey(RenderObject* pRenderObject, const Vector3F& rPosition);
 };
 
 }
